Add option -r | --remove to specify checks to remove from the report, it can be used multiple time. The value must be the number of a check or a regexp.
To allow this check's exclusion all checks are now done in a separate
funtion named with the number of the check. For example check 3.2 is
done in function check_3_2 and check 3.1.17 in method check_3_1_17.
All section number not related to a check must also have a function
defined that is doing nothing, see check_4 for example.